+++ practice/core/utils/play_utils.py
@@ -10,8 +10,8 @@ from torch import nn
 
 from practice.core.base.config import BaseConfig
 from practice.core.base.context import ContextBase
+from practice.core.utils.eval_utils import get_action
 from practice.exercise1_q.q_table_exercise import QTable
-from practice.utils.eval_utils import get_action
 from practice.utils_for_coding.network_utils import load_model
