+++ practice/core/utils/tests/test_hub.py
@@ -28,7 +28,7 @@ import torch
 
 from practice.core.base.config import ArtifactConfig, BaseConfig, EnvConfig
 from practice.core.base.trainer import TrainerBase
-from practice.utils.hub_utils import push_model_to_hub, push_to_hub_generic
+from practice.core.utils.hub_utils import push_model_to_hub, push_to_hub_generic
 
 
 @pytest.fixture
@@ -118,8 +118,8 @@ def mock_metadata() -> dict[str, Any]:
 class TestPushToHubGeneric:
     """Test push_to_hub_generic function."""
 
-    @patch("practice.utils.hub_utils.push_model_to_hub")
-    @patch("practice.utils.hub_utils._get_env_name_and_metadata")
+    @patch("practice.core.utils.hub_utils.push_model_to_hub")
+    @patch("practice.core.utils.hub_utils._get_env_name_and_metadata")
     def test_push_to_hub_success(
         self,
         mock_get_metadata: Mock,
@@ -165,8 +165,8 @@ class TestPushToHubGeneric:
         assert test_config.artifact_config.usage_instructions in model_card
         assert test_config.artifact_config.model_filename in model_card
 
-    @patch("practice.utils.hub_utils.push_model_to_hub")
-    @patch("practice.utils.hub_utils._get_env_name_and_metadata")
+    @patch("practice.core.utils.hub_utils.push_model_to_hub")
+    @patch("practice.core.utils.hub_utils._get_env_name_and_metadata")
     def test_push_to_hub_with_different_username(
         self,
         mock_get_metadata: Mock,
@@ -189,8 +189,8 @@ class TestPushToHubGeneric:
         call_args = mock_push_model.call_args
         assert call_args[1]["repo_id"] == f"{username}/{test_config.artifact_config.repo_id}"
 
-    @patch("practice.utils.hub_utils.push_model_to_hub")
-    @patch("practice.utils.hub_utils._get_env_name_and_metadata")
+    @patch("practice.core.utils.hub_utils.push_model_to_hub")
+    @patch("practice.core.utils.hub_utils._get_env_name_and_metadata")
     def test_push_to_hub_with_no_extra_tags(
         self,
         mock_get_metadata: Mock,
@@ -224,8 +224,8 @@ class TestPushToHubGeneric:
             extra_tags=(),
         )
 
-    @patch("practice.utils.hub_utils.push_model_to_hub")
-    @patch("practice.utils.hub_utils._get_env_name_and_metadata")
+    @patch("practice.core.utils.hub_utils.push_model_to_hub")
+    @patch("practice.core.utils.hub_utils._get_env_name_and_metadata")
     def test_push_to_hub_metadata_error(
         self,
         mock_get_metadata: Mock,
@@ -246,8 +246,8 @@ class TestPushToHubGeneric:
         # NOTE: Environment cleanup is now handled by the CLI, not by push_to_hub_generic
         # So we don't check for env.close() being called here
 
-    @patch("practice.utils.hub_utils.push_model_to_hub")
-    @patch("practice.utils.hub_utils._get_env_name_and_metadata")
+    @patch("practice.core.utils.hub_utils.push_model_to_hub")
+    @patch("practice.core.utils.hub_utils._get_env_name_and_metadata")
     def test_push_to_hub_push_error(
         self,
         mock_get_metadata: Mock,
@@ -270,8 +270,8 @@ class TestPushToHubGeneric:
         # NOTE: Environment cleanup is now handled by the CLI, not by push_to_hub_generic
         # So we don't check for env.close() being called here
 
-    @patch("practice.utils.hub_utils.push_model_to_hub")
-    @patch("practice.utils.hub_utils._get_env_name_and_metadata")
+    @patch("practice.core.utils.hub_utils.push_model_to_hub")
+    @patch("practice.core.utils.hub_utils._get_env_name_and_metadata")
     def test_model_card_generation(
         self,
         mock_get_metadata: Mock,
@@ -306,7 +306,7 @@ class TestPushToHubGeneric:
 class TestPushModelToHub:
     """Test push_model_to_hub function."""
 
-    @patch("practice.utils.hub_utils._push_to_hub")
+    @patch("practice.core.utils.hub_utils._push_to_hub")
     def test_push_model_to_hub_success(
         self, mock_push_to_hub: Mock, test_config: BaseConfig, mock_metadata: dict[str, Any]
     ) -> None:
@@ -355,7 +355,7 @@ class TestPushModelToHub:
         ]
         assert call_args[1]["file_pathnames"] == expected_files
 
-    @patch("practice.utils.hub_utils._push_to_hub")
+    @patch("practice.core.utils.hub_utils._push_to_hub")
     def test_push_model_to_hub_with_different_files(
         self, mock_push_to_hub: Mock, mock_metadata: dict[str, Any], temp_output_dir: Path
     ) -> None:
@@ -410,7 +410,7 @@ class TestPushModelToHub:
         assert call_args[1]["file_pathnames"] == expected_files
         assert call_args[1]["eval_result_pathname"] == str(temp_output_dir / "custom_eval.json")
 
-    @patch("practice.utils.hub_utils._push_to_hub")
+    @patch("practice.core.utils.hub_utils._push_to_hub")
     def test_push_model_to_hub_error(
         self, mock_push_to_hub: Mock, test_config: BaseConfig, mock_metadata: dict[str, Any]
     ) -> None:
@@ -451,8 +451,8 @@ class TestPushModelToHub:
 class TestHubUtilsIntegration:
     """Integration tests for hub utilities."""
 
-    @patch("practice.utils.hub_utils._push_to_hub")
-    @patch("practice.utils.hub_utils._get_env_name_and_metadata")
+    @patch("practice.core.utils.hub_utils._push_to_hub")
+    @patch("practice.core.utils.hub_utils._get_env_name_and_metadata")
     def test_full_hub_workflow(
         self,
         mock_get_metadata: Mock,
