+++ practice/exercise10_ddp_ppo/ppo_rnd_exercise.py
@@ -13,8 +13,8 @@ from practice.core.base.context import ContextBase
 from practice.core.base.env_typing import ActTypeC, ObsType
 from practice.core.base.rewarder import RewardBase, RewardConfig
 from practice.core.base.trainer import TrainerBase
+from practice.core.utils.dist_utils import unwrap_model
 from practice.exercise7_ppo.ppo_exercise import _RolloutBuffer, _StepData
-from practice.utils.dist_utils import unwrap_model
 from practice.utils_for_coding.network_utils import MLP, LogStdHead, init_weights
 from practice.utils_for_coding.numpy_tensor_utils import np2tensor, tensor2np_1d
 from practice.utils_for_coding.rollout_utils import get_good_transition_mask
