+++ practice/exercise1_q/tests/test_q_table_train.py
@@ -33,11 +33,11 @@ from typing import Generator
 import numpy as np
 import pytest
 
+from practice.core.utils.play_utils import play_and_generate_video_generic
+from practice.core.utils.train_utils import train_and_evaluate_network
 from practice.exercise1_q.config_taxi import generate_context
 from practice.exercise1_q.q_table_exercise import QTable
 from practice.exercise1_q.q_trainer_exercise import QTableConfig, QTableTrainer
-from practice.utils.play_utils import play_and_generate_video_generic
-from practice.utils.train_utils import train_and_evaluate_network
 from practice.utils_for_coding.scheduler_utils import ExponentialSchedule
 
 
@@ -45,7 +45,7 @@ from practice.utils_for_coding.scheduler_utils import ExponentialSchedule
 def test_config() -> QTableConfig:
     """Create a test configuration based on taxi config with reduced parameters."""
     from practice.core.base.config import ArtifactConfig, EnvConfig
-    from practice.utils.env_utils import get_device
+    from practice.core.utils.env_utils import get_device
 
     device = get_device()
