+++ practice/exercise2_dqn/e24_rainbow/config_lunar_2d.py
@@ -3,11 +3,11 @@ from torch.optim import Adam
 
 from practice.core.base.config import ArtifactConfig, EnvConfig
 from practice.core.base.context import ContextBase
+from practice.core.utils.env_utils import get_device, get_env_from_config
 from practice.exercise2_dqn.dqn_trainer import DQNTrainer
 from practice.exercise2_dqn.e24_rainbow.config_lunar_1d import schedule_beta_increment
 from practice.exercise2_dqn.e24_rainbow.per_exercise import PERBufferConfig
 from practice.exercise2_dqn.e24_rainbow.rainbow_exercise import RainbowConfig, RainbowNet
-from practice.utils.env_utils import get_device, get_env_from_config
 from practice.utils_for_coding.network_utils import load_checkpoint_if_exists
 from practice.utils_for_coding.scheduler_utils import ConstantSchedule
