+++ practice/exercise3_reinforce/tests/test_reinforce_train.py
@@ -31,14 +31,14 @@ import pytest
 import torch
 
 from practice.core.base.config import ArtifactConfig, EnvConfig
+from practice.core.utils.env_utils import get_device
+from practice.core.utils.train_utils import train_and_evaluate_network
 from practice.exercise3_reinforce.config_cartpole import generate_context
 from practice.exercise3_reinforce.reinforce_exercise import (
     Reinforce1DNet,
     ReinforceConfig,
     ReinforceTrainer,
 )
-from practice.utils.env_utils import get_device
-from practice.utils.train_utils import train_and_evaluate_network
 from practice.utils_for_coding.scheduler_utils import ConstantSchedule
