+++ practice/exercise4_curiosity/config_mountain_car.py
@@ -3,6 +3,7 @@ import torch.optim as optim
 from gymnasium.spaces import Discrete
 
 from practice.core.base.config import ArtifactConfig, EnvConfig
+from practice.core.utils.env_utils import get_device, get_env_from_config
 from practice.exercise3_reinforce.reinforce_exercise import Reinforce1DNet
 from practice.exercise4_curiosity.curiosity_exercise import (
     RND1DNetworkConfig,
@@ -14,7 +15,6 @@ from practice.exercise4_curiosity.enhanced_reinforce import (
     EnhancedReinforceTrainer,
     ReinforceContext,
 )
-from practice.utils.env_utils import get_device, get_env_from_config
 from practice.utils_for_coding.baseline_utils import ConstantBaseline
 from practice.utils_for_coding.network_utils import load_checkpoint_if_exists
 from practice.utils_for_coding.scheduler_utils import LinearSchedule
