+++ practice/exercise5_a2c/tests/test_a2c_train.py
@@ -5,9 +5,9 @@ from typing import Generator
 
 import pytest
 
+from practice.core.utils.train_utils import train_and_evaluate_network
 from practice.exercise5_a2c.a2c_gae_exercise import A2CConfig
 from practice.exercise5_a2c.config_cartpole import generate_context, get_app_config
-from practice.utils.train_utils import train_and_evaluate_network
 
 
 @pytest.fixture
