+++ practice/exercise8_td3/config_pendulum.py
@@ -4,16 +4,16 @@ import torch.optim as optim
 
 from practice.core.base.config import ArtifactConfig, EnvConfig
 from practice.core.base.env_typing import EnvsTypeC, EnvTypeC
-from practice.exercise8_td3.td3_exercise import (
-    TD3Actor,
-    TD3Config,
-    TD3Trainer,
-)
-from practice.utils.env_utils import (
+from practice.core.utils.env_utils import (
     get_device,
     get_env_from_config,
     verify_env_with_continuous_action,
     verify_vector_env_with_continuous_action,
+)
+from practice.exercise8_td3.td3_exercise import (
+    TD3Actor,
+    TD3Config,
+    TD3Trainer,
 )
 from practice.utils_for_coding.context_utils import ACContext
 from practice.utils_for_coding.network_utils import DoubleQCritic
