+++ practice/exercise8_td3/tests/test_td3_train.py
@@ -5,9 +5,9 @@ from typing import Generator
 
 import pytest
 
+from practice.core.utils.train_utils import train_and_evaluate_network
 from practice.exercise8_td3.config_pendulum import generate_context, get_app_config
 from practice.exercise8_td3.td3_exercise import TD3Config
-from practice.utils.train_utils import train_and_evaluate_network
 
 
 @pytest.fixture
