+++ practice/exercise9_sac/tests/test_sac_train.py
@@ -5,9 +5,9 @@ from typing import Generator
 
 import pytest
 
+from practice.core.utils.train_utils import train_and_evaluate_network
 from practice.exercise9_sac.config_pendulum import generate_context, get_app_config
 from practice.exercise9_sac.sac_exercise import SACConfig
-from practice.utils.train_utils import train_and_evaluate_network
 
 
 @pytest.fixture
